Figured out listening on all ip addresses.
Windows is pretty strict on how it allows binding to ip addresses for http services to listen for things.  Had to add a urlacl rule using netsh http settings from an admin level command prompt.  This setting will persist across reboots.  But because we have to allow binding to a specific port from the netsh tool, we will hard code that port in the dll for simplicity sake.  Future versions should allow a more flexible approach.